+++ CHANGELOG.md
@@ -5,6 +5,7 @@ IMPROVEMENTS:
 * resource/nomad_namespace: add `vault_config` and `consul_config` blocks to configure Vault and Consul cluster permissions (Nomad Enterprise only) ([#581](https://github.com/hashicorp/terraform-provider-nomad/pull/581))
 * data source/nomad_node_pool: Added the `node_identity_ttl` argument ([#569](https://github.com/hashicorp/terraform-provider-nomad/pull/569))
 * resource/nomad_node_pool: Added the `node_identity_ttl` argument to configure the node identity TTL for nodes in the pool ([#569](https://github.com/hashicorp/terraform-provider-nomad/pull/569))
+* resource/nomad_quota_specification: add support for all `QuotaResources` fields including `secrets_mb`, `devices`, `numa`, and `storage` ([#584](https://github.com/hashicorp/terraform-provider-nomad/pull/584))
 
 ## 2.5.2 (November 13, 2025)